+++ RayTracer/src/vertex.h
@@ -2,14 +2,14 @@
 #define VERTEX_H_
 
 struct Vertex {
-  float x;   // x coordinate
-  float y;   // y coordinate
-  float z;   // z coordinate
-  float nx;  // x component of normal
-  float ny;  // y component of normal
-  float nz;  // z component of normal
-  float u;   // u (or s) coordinate of texture
-  float v;   // v (or t) coordinate of texture
+  double x;   // x coordinate
+  double y;   // y coordinate
+  double z;   // z coordinate
+  double nx;  // x component of normal
+  double ny;  // y component of normal
+  double nz;  // z component of normal
+  double u;   // u (or s) coordinate of texture
+  double v;   // v (or t) coordinate of texture
 };
 
 #endif
